Transform Movebank data to a Darwin Core Archive

Transforms a Movebank dataset (formatted as a Frictionless Data Package) to a Darwin Core Archive.

Usage

write_dwc(package, directory = ".", doi = package$id, rights_holder = NULL)

Arguments

package

A Frictionless Data Package of Movebank data, as read by frictionless::read_package(). It is expected to contain a reference-data and gps resource.

directory

Path to local directory to write file to.

doi

DOI of the original dataset, used to get dataset-level terms.

rights_holder

Acronym of the organization owning or managing the rights over the data.

Value

CSV and meta.xml files written to disk. And invisibly, a list of data frames with the transformed data.

Details

The resulting files can be uploaded to an IPT for publication to GBIF and/or OBIS. A corresponding eml.xml metadata file can be created with write_eml(). See Get started for examples.

Transformation details

This function follows recommendations suggested by Peter Desmet, Sarah Davidson, John Wieczorek and others and transforms data to:

An Occurrence core.

A meta.xml file.

Key features of the Darwin Core transformation:

Deployments (animal+tag associations) are parent events, with tag attachment (a human observation) and GPS positions (machine observations) as child events. No information about the parent event is provided other than its ID, meaning that data can be expressed in an Occurrence core with one row per observation and parentEventID shared by all occurrences in a deployment.

The tag attachment event often contains metadata about the animal (sex, life stage, comments) and deployment as a whole.

No event/occurrence is created for the deployment end, since the end date is often undefined, unreliable and/or does not represent an animal occurrence.

Only visible (non-outlier) GPS records that fall within a deployment are included.

GPS positions are downsampled to the first GPS position per hour, to reduce the size of high-frequency data. It is possible for a deployment to contain no GPS positions, e.g. if the tag malfunctioned right after deployment.

See also

Examples

write_dwc(o_assen, directory = "my_directory")
#> 
#> ── Reading data ──
#> 
#> ℹ Taxa found in reference data and their WoRMS AphiaID:
#> Haematopus ostralegus: 147436
#> 
#> ── Transforming data to Darwin Core ──
#> 
#> ── Writing files ──
#> 
#> • my_directory/dwc_occurrence.csv

# Clean up (don't do this if you want to keep your files)
unlink("my_directory", recursive = TRUE)